add holo

+++ rust/autd3/defined.rs
@@ -4,7 +4,7 @@
  * Created Date: 02/05/2022
  * Author: Shun Suzuki
  * -----
- * Last Modified: 28/07/2022
+ * Last Modified: 11/05/2023
  * Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
  * -----
  * Copyright (c) 2022 Shun Suzuki. All rights reserved.
@@ -16,9 +16,9 @@ use num::FromPrimitive;
 pub const NUM_TRANS_IN_UNIT: usize = 249;
 pub const NUM_TRANS_X: usize = 18;
 pub const NUM_TRANS_Y: usize = 14;
-pub const TRANS_SPACING_MM: f64 = 10.16;
-pub const DEVICE_WIDTH: f64 = 192.0;
-pub const DEVICE_HEIGHT: f64 = 151.4;
+pub const TRANS_SPACING_MM: float = 10.16;
+pub const DEVICE_WIDTH: float = 192.0;
+pub const DEVICE_HEIGHT: float = 151.4;
 
 pub fn is_missing_transducer<T1, T2>(x: T1, y: T2) -> bool
 where
